Issue #12: Extracted all cast methods from `castHelpers` to `cast` module.
