+++ bot/cast/__init__.py
@@ -1,0 +1,11 @@
+from .replyOrSend import replyOrSend
+from .sendInfoToChat import sendInfoToChat
+from .downloadAndSendAudioToChat import downloadAndSendAudioToChat
+from . import castConfig
+
+__all__ = [
+    'replyOrSend',
+    'sendInfoToChat',
+    'downloadAndSendAudioToChat',
+    'castConfig',
+]
