+++ bot/cast/getYtdlBaseOptions.py
@@ -1,0 +1,43 @@
+# -*- coding:utf-8 -*-
+
+from core.appConfig import appConfig, TEMP_PATH
+from core.logger import getLogger
+
+from .YtdlOptionsType import YtdlOptionsType
+
+_logger = getLogger('bot/cast/getYtdlBaseOptions')
+
+
+def getYtdlBaseOptions():
+    # Prepare options:
+    options: YtdlOptionsType = {
+        # @see https://github.com/ytdl-org/youtube-dl/blob/3e4cedf9e8cd3157df2457df7274d0c842421945/youtube_dl/YoutubeDL.py#L137-L312
+        'verbose': True,
+        'cachedir': TEMP_PATH,
+        'verbose': True,
+        'noplaylist': True,
+        'keepvideo': False,
+        'extractor_args': None,
+        'outtmpl': None,
+        '_destFolder': None,
+        '_destFile': None,
+        'cookiefile': None,
+        'format': None,
+        #  'debug_printtraffic': None,
+    }
+
+    # Add PO Token (if exists), see https://github.com/yt-dlp/yt-dlp/wiki/Extractors#manually-acquiring-a-po-token-from-a-browser-for-use-when-logged-out
+    YT_POTOKEN = appConfig.get('YT_POTOKEN')
+    if YT_POTOKEN:
+        _logger.info('getYtdlBaseOptions: Using YT_POTOKEN: %s' % (YT_POTOKEN))
+        options['extractor_args'] = 'youtube:player-client=web;po_token=web+' + YT_POTOKEN
+
+    #  # Add authentication params (NOTE: Unused as not supported)
+    #  YT_USERNAME = appConfig.get('YT_USERNAME')
+    #  YT_PASSWORD = appConfig.get('YT_PASSWORD')
+    #  if YT_USERNAME and YT_PASSWORD:
+    #      _logger.info('getYtdlBaseOptions: Using username (%s) and password (***)' % (YT_USERNAME))
+    #      options['username'] = YT_USERNAME
+    #      options['password'] = YT_PASSWORD
+
+    return options
