+++ bot/commands/castTestCommand.py
@@ -4,11 +4,11 @@ import telebot  # pyTelegramBotAPI
 
 from core.logger import getLogger
 from bot.botApp import botApp
-
 from core.appConfig import TELEGRAM_OWNER_ID
 
+from bot.cast import downloadAndSendAudioToChat
+from bot.cast.castConfig import demoVideo
 
-from .castHelpers import demoVideo, downloadAndSendAudioToChat
 
 _logger = getLogger('bot/commands/castTestCommand')
