+++ core/appConfig.py
@@ -54,8 +54,11 @@ TELEGRAM_OWNER_ID = int(appConfig.get('TELEGRAM_OWNER_ID', '0'))
 VERCEL_URL = str(appConfig.get('VERCEL_URL', ''))
 IS_VERCEL = True if VERCEL_URL else False
 
+# Current root project path
+CWD_PATH = pathlib.Path(os.getcwd()).as_posix()
+
 # Temp path: Use local 'temp' or vercel specific '/tmp' folders for temporary files
-TEMP_PATH = posixpath.join(pathlib.Path(os.getcwd()).as_posix(), 'temp') if LOCAL or not IS_VERCEL else '/tmp'
+TEMP_PATH = posixpath.join(CWD_PATH, 'temp') if LOCAL or not IS_VERCEL else '/tmp'
 
 # Audio...
